requirements.txt
streamlit==1.24.1
requests==2.29.0
openai==0.27.4
pillow==9.4.0